E2

+++ web/packages/scene-runtime/src/terrain-foliage-shader.ts
@@ -112,7 +112,7 @@ const createTerrainFoliageShaderEffect = (id: string): RenderShaderEffectDefinit
         ],
     },
     fragment: {
-        precision: 'highp',
+        precision: 'mediump',
         inputs: [
             { name: 'v_Normal', type: 'vec3' },
             { name: 'v_UV', type: 'vec2' },
